requirements.txt
dspy-ai
google-genai
python-dotenv